boot: vendor specific location for systemd-boot
No matter how we twist and turn it, systemd-boot is going to be
vendor specific one way or another. Especially in secure-boot
context it has to have vendor specific SBAT values and a vendor
specific signature. It may also have to be installed alongside a vendor
specific shim.
So it makes sense for operating system vendors to put systemd-boot into
/EFI/$VENDOR instead of the generic /EFI/systemd. The latter could still
be used by self compiled versions or upstream builds.